soc-final/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=3.0.1
groq>=0.9.0
numpy>=2.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
netmiko>=4.4.0
scikit-learn>=1.5.2

# Retrieval / RAG dependencies
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
